Break out `UploadManagerContext` into its own file
