+++ src/components/Upload/index.ts
@@ -1,4 +1,5 @@
 export * from './Reducer';
 export * from './UploadContext';
+export * from './UploadManagerContext';
 export * from './UploadFilesForm';
 export * from './UploadManager';
